Stash Settings: Add session support
